changed file structure. moved bottom sheet vc logic to different class
