+++ iBench/UI/Views/Extensions/UIActivityIndicatorView+IsActive.swift
